Tab Bar and constraints (some screens)
Implemented a 3-tab tab bar, added a non-scrollable header to the transactions table, added SwiftLint to the project, added a file for UI specifications that I can use to declare custom components (e.g., colour) that I would like to use in the application. Modified the logic for signing up and logging in.